refactor - split analyzers into different files
